requirements.txt
pandas
openpyxl
python-dotenv
yagmail
pytest
